requirements.txt
Django>=6.0,<6.1
dj-database-url>=2.3,<3.0
django-storages[google]>=1.14.6,<2.0
gunicorn>=23.0,<24.0
google-api-python-client>=2.170,<3.0
google-auth>=2.40,<3.0
google-auth-oauthlib>=1.2,<2.0
psycopg2-binary>=2.9,<3.0
whitenoise[brotli]>=6.9,<7.0
